requirements.txt
pandas>=2.0
numpy>=1.24
python-binance>=1.0.19
ta>=0.11.0
matplotlib>=3.7
plotly>=5.15
tabulate>=0.9
tqdm>=4.65
python-dotenv>=1.0
pytest>=7.0
pyarrow>=14.0
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
